- Creates Rooms model - Adds seeder for Rooms.

+++ database/migrations/2019_09_11_134908_create_rooms_table.php
@@ -15,8 +15,8 @@ class CreateRoomsTable extends Migration
     {
         Schema::create('rooms', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedTinyInteger('floor');
-            $table->unsignedTinyInteger('room_number');
+            $table->unsignedInteger('floor');
+            $table->unsignedInteger('room_number');
             $table->string('description', 200)->nullable();
             $table->boolean('out_of_order')->default(false)->nullable(false);
             $table->timestamps();
